% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Exposure Calculation for Clustered Interference}
\usage{
exposure(z, K)
}
\arguments{
\item{z}{A numeric vector representing treatment assignments.}

\item{K}{An integer specifying the number of clusters.}
}
\value{
A numeric vector representing the exposure of each unit.
}
\description{
This function constructs an exposure vector based on the assignment vector z
and the number of clusters K.
}
