% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_normal.R
\name{Imputation_normal}
\alias{Imputation_normal}
\title{Imputation with Normal Distribution}
\usage{
Imputation_normal(null, exposure, observation)
}
\arguments{
\item{null}{A vector of two values representing two exposure groups specified
in the null hypothesis.
It should specify the two levels of the 'exposure' variable that are being
compared.}

\item{exposure}{A vector indicating the exposure status for each observation.
The values in this vector should contain the levels provided in the 'null'
vector.}

\item{observation}{A numeric vector of observed values.}
}
\value{
A vector of imputed values sampled from the observed data.
}
\description{
This function performs imputation for missing observations based on observed
data.
The observed data are modeled with a normal distribution, and
the missing observations are imputed with the posterior predictive
distribution
with non-informative prior.
}
