% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_sample.R
\name{Imputation_sample}
\alias{Imputation_sample}
\title{Imputation with Empirical Distribution of Observed Outcomes}
\usage{
Imputation_sample(null, exposure, observation)
}
\arguments{
\item{null}{A vector of two values representing two exposure groups specified
in the null hypothesis.
It should specify the two levels of the 'exposure' variable that are being
compared.}

\item{exposure}{A vector indicating the exposure status for each observation.
The values in this vector should contain the levels provided in the 'null'
vector.}

\item{observation}{A numeric vector of observed values.}
}
\value{
A vector of imputed values sampled from the observed data.
}
\description{
This function performs imputation for missing observations
by sampling from the observed data.
}
