% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{test_statistic}
\alias{test_statistic}
\title{Compute Test Statistic}
\usage{
test_statistic(null, exposure, observation)
}
\arguments{
\item{null}{A vector of two values representing two exposure groups
specified in the null hypothesis.
It should specify the two levels of the 'exposure' variable that are being
compared.}

\item{exposure}{A vector indicating the exposure status for each observation.
The values in this vector should contain the levels provided in the 'null'
vector.}

\item{observation}{A numeric vector of observed values. The function will
compute the means
of this vector for the two exposure groups and return the absolute
difference.}
}
\value{
The absolute difference between the means of the 'observation'
values for the two
exposure groups specified in 'null'.
}
\description{
This function calculates the absolute difference between the means of
the observed outcomes for the two groups specified in the null hypothesis.
}
